Allow #AUTODOC# line to be anywhere in the file
Skip comments that are clearly not part of a description
Ie, we can have multiline descriptions in our autodoc comments, but they must be consecutive. If we have a comment line followed by another comment line, they're a single multi-line description. If however, 10 lines later (without any comments inbetween) we have another comment line, assume that's just a code comment, not a continuation of the description